cleanup(virtctl,create): Move volume type definitions into separate file
Move the various volume type definitions into a separate file to make
top level variable definitions in vm.go more readable.
