+++ pkg/virtctl/create/vm/BUILD.bazel
@@ -2,7 +2,10 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = ["vm.go"],
+    srcs = [
+        "vm.go",
+        "volume.go",
+    ],
     importpath = "kubevirt.io/kubevirt/pkg/virtctl/create/vm",
     visibility = ["//visibility:public"],
     deps = [
