+++ pkg/virtctl/create/vm/vm.go
@@ -106,30 +106,6 @@ type createVM struct {
 	memoryChanged                 bool
 }
 
-type cloneVolume struct {
-	Name      string             `param:"name"`
-	Source    string             `param:"src"`
-	BootOrder *uint              `param:"bootorder"`
-	Size      *resource.Quantity `param:"size"`
-}
-
-type containerdiskVolume struct {
-	Name      string `param:"name"`
-	Source    string `param:"src"`
-	BootOrder *uint  `param:"bootorder"`
-}
-
-type pvcVolume struct {
-	Name      string `param:"name"`
-	Source    string `param:"src"`
-	BootOrder *uint  `param:"bootorder"`
-}
-
-type blankVolume struct {
-	Name string             `param:"name"`
-	Size *resource.Quantity `param:"size"`
-}
-
 var optFns = map[string]func(*createVM, *v1.VirtualMachine) error{
 	RunStrategyFlag:          withRunStrategy,
 	InstancetypeFlag:         withInstancetype,
@@ -160,87 +136,6 @@ var flags = []string{
 	VolumeImportFlag,
 	CloudInitUserDataFlag,
 	CloudInitNetworkDataFlag,
-}
-
-type dataVolumeSourceBlank struct {
-	Size *resource.Quantity `param:"size"`
-	Type string             `param:"type"`
-	Name string             `param:"name"`
-}
-
-type dataVolumeSourceGcs struct {
-	SecretRef string             `param:"secretref"`
-	URL       string             `param:"url"`
-	Size      *resource.Quantity `param:"size"`
-	Type      string             `param:"type"`
-	Name      string             `param:"name"`
-}
-
-type dataVolumeSourceHttp struct {
-	CertConfigMap      string             `param:"certconfigmap"`
-	ExtraHeaders       []string           `param:"extraheaders"`
-	SecretExtraHeaders []string           `param:"secretextraheaders"`
-	SecretRef          string             `param:"secretref"`
-	URL                string             `param:"url"`
-	Size               *resource.Quantity `param:"size"`
-	Type               string             `param:"type"`
-	Name               string             `param:"name"`
-}
-
-type dataVolumeSourceImageIO struct {
-	CertConfigMap string             `param:"certconfigmap"`
-	DiskId        string             `param:"diskid"`
-	SecretRef     string             `param:"secretref"`
-	URL           string             `param:"url"`
-	Size          *resource.Quantity `param:"size"`
-	Type          string             `param:"type"`
-	Name          string             `param:"name"`
-}
-
-type dataVolumeSourcePVC struct {
-	Name   string             `param:"name"`
-	Source string             `param:"src"`
-	Size   *resource.Quantity `param:"size"`
-	Type   string             `param:"type"`
-}
-
-type dataVolumeSourceRegistry struct {
-	CertConfigMap string             `param:"certconfigmap"`
-	ImageStream   string             `param:"imagestream"`
-	PullMethod    string             `param:"pullmethod"`
-	SecretRef     string             `param:"secretref"`
-	URL           string             `param:"url"`
-	Size          *resource.Quantity `param:"size"`
-	Type          string             `param:"type"`
-	Name          string             `param:"name"`
-}
-
-type dataVolumeSourceS3 struct {
-	CertConfigMap string             `param:"certconfigmap"`
-	SecretRef     string             `param:"secretref"`
-	URL           string             `param:"url"`
-	Size          *resource.Quantity `param:"size"`
-	Type          string             `param:"type"`
-	Name          string             `param:"name"`
-}
-
-type dataVolumeSourceVDDK struct {
-	BackingFile  string             `param:"backingfile"`
-	InitImageUrl string             `param:"initimageurl"`
-	SecretRef    string             `param:"secretref"`
-	ThumbPrint   string             `param:"thumbprint"`
-	URL          string             `param:"url"`
-	UUID         string             `param:"uuid"`
-	Size         *resource.Quantity `param:"size"`
-	Type         string             `param:"type"`
-	Name         string             `param:"name"`
-}
-
-type dataVolumeSourceSnapshot struct {
-	Name   string             `param:"name"`
-	Source string             `param:"src"`
-	Size   *resource.Quantity `param:"size"`
-	Type   string             `param:"type"`
 }
 
 var volumeImportOptions = map[string]func(string) (*cdiv1.DataVolumeSource, error){
